Cycle fixes: extract event kind types and utils
